requirements.txt
Django==4.2.4
djangorestframework==3.14.0
django-filter==23.3
psycopg2-binary
pillow